% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imnr-utils.r
\name{add.snr.ref.limits}
\alias{add.snr.ref.limits}
\title{Draw SNR reference limits}
\usage{
add.snr.ref.limits(p)
}
\arguments{
\item{p}{A valid ggplot2 object}
}
\description{
Draw in a SNR (in dB) plot made with the ggplot2 package SNR reference
quality limits.
}
\details{
The red limit marks the 'acceptable' limit
the yellow limit marks the 'excelent' limit
the yellow limit marks the 'excelent' limit
}

