% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{adobeRGB.from.L.Lab}
\alias{adobeRGB.from.L.Lab}
\title{Get Adobe RGB tonal values from L.Lab ones}
\usage{
adobeRGB.from.L.Lab(L)
}
\arguments{
\item{Adobe}{RGB tonal values in \code{[0, 1]}}
}
\value{
L.Lab tonal values in \code{[0, 100]}
}
\description{
Convert Lightness (L from Lab) tonal values to Adobe RGB ones.
}

