% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{adobeRGB.from.linear}
\alias{adobeRGB.from.linear}
\title{Get Adobe RGB tonal values from linear ones}
\usage{
adobeRGB.from.linear(linear)
}
\arguments{
\item{linear}{tonal values in \code{[0, 1]}}
}
\value{
Adobe RGB tonal values in \code{[0, 1]}
}
\description{
Convert linear tonal values to Adobe RGB ones.
}

