% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{adobeRGB.from.sRGB}
\alias{adobeRGB.from.sRGB}
\title{Get Adobe RGB tonal values from sRGB ones}
\usage{
adobeRGB.from.sRGB(sRGB)
}
\arguments{
\item{sRGB}{tonal values in \code{[0, 1]}}
}
\value{
Adobe RGB tonal values in \code{[0, 1]}
}
\description{
Convert sRGB tonal values to Adobe RGB ones.
}

