% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tone-curves.r
\name{build.generic.gamma.curve}
\alias{build.generic.gamma.curve}
\title{}
\usage{
build.generic.gamma.curve(gamma, gamma.slope, gamma.intercept, linear.slope,
  input.max.linear, num.of.linear.points, step)
}
\description{

}

