% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imnr-utils.r
\name{file.name.seq}
\alias{file.name.seq}
\title{Prepare a sequence of file names}
\usage{
file.name.seq(file.name.base, file.count, file.name.ext)
}
\arguments{
\item{file.name.base}{The base file name, a constant prefix of the file
  finames.}

\item{file.count}{The number of files in the sequence. The returning list
  will contain a file with file counter from 1 to up to this argument.}

\item{file.name.ext}{The file name extension. A constant for all the file
  names.}
}
\value{
A character vector with \code{img.file.count} elements, with the
  pattern <file.name.base><file.counter>.<file.name.ext>. Where
  <file.counter> will vary from 1 to \code{file.count}
}
\description{
Return a character vector with file names following a sequence pattern.
}
\details{
This is a convenience function for the case when you want to prepare a
vector of file names with the following pattern:

<file.name.base><file.counter>.<file.name.ext>

As in ('crop_1.fit', 'crop_2.fit', ... 'crop_75.fit')
}

