% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noiseVar_class.R
\name{hvdvm$channel.labels}
\alias{hvdvm$channel.labels}
\alias{vvm$channel.labels}
\title{Get the channel labels}
\usage{
hvdvm$channel.labels

  vvm$channel.labels

hvdvm$model.list

  vvm$model.list
}
\value{
A character vector with five values. The first two are for the first
  row in the 2x2 Bayer pattern, the second pair are for the second row, and
  the fifth to label computations merging both green channels.

A list where the names are those given to the fitted models and the
  values are a character value with the call used to fit the model.
}
\description{
A vector with the channel labels. This labels are set when an instance of the
class is constructed using the \code{new} function.

Get a list with the currently fitted models.
}
\examples{
\dontrun{

# Print the channel labels
my.hvdvm$channel.labels
#> [1] "Blue"    "Green B" "Green R" "Red"
}
\dontrun{

# Print the list of models
my.hvdvm$model.list
#> $standard
#> [1] "lm(formula = var ~ mean, data = model.src('std.var'))"
#>
#> $weighted
#> [1] "lmrob(formula = var ~ mean, data = model.src('std.var'), weights = 1/mean^2)"
}
}
\seealso{
\code{\link{hvdvm$new}}, \code{\link{vvm$new}}

\code{\link{hvdvm$fit.model}}, \code{\link{vvm$fit.model}}
}

