% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hvdvm-class.r
\name{hvdvm$cov.df}
\alias{hvdvm$cov.df}
\title{Get the resulting covariance data.}
\value{
A data.frame with the covariance data from the channels in each delta
  image computed by the \code{\link{hvdvm$digest}} function. This data frame
  contains the following columns:

\itemize{

 \item \code{cond}: A key for the photographic conditions common to the two
 processed samples (\code{pict1, pict2}). The detail of the photographic
 conditions is returned by the \code{\link{hvdvm$photo.conditions.df}}
 function, where the key is the value in this \code{cond} value.

 \item \code{pict1, pict2}: Name of the processed image samples files.

 \item \code{chan.a, chan.b}: The label of the channels whose covariance is shown
 in the \code{cov} column.

 \item \code{cov}: The half of the photosite values covariance in the
 \code{chan.a} and \code{chan.b} channels in the difference image, computed
 pixel by pixel, of the \code{pict1} image minus the \code{pict2} one.
}
}
\description{
Returns the covariance data computed by the \code{\link{hvdvm$digest}}
function. This values are half of the covariance in the delta image, computed
for each pair of channels combination.
}
\section{Usage}{

 \preformatted{
  hvdvm$cov.df
 }
}
\examples{
\dontrun{

# print the head of the variance data frame
head(my.hvdvm$cov.df)

# Merge the variance and covariance data
merged <- merge(hvd$var.df, hvd$cov.df, by=c('pict1','pict2', 'cond'))
# Plot all the covariances (y axis) with respect to the means (x axis)
plot(merged$mean, merged$cov)
}
}
\seealso{
\code{\link{hvdvm$var.df}}, \code{\link{hvdvm$photo.conditions.df}}
}

