% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/hvdvm_class.R
\name{hvdvm$digest}
\alias{hvdvm$digest}
\title{Process the image samples.}
\usage{
hvdvm$digest(
     photo.conds.file,
     crop.files.path = './'
     )
}
\arguments{
\item{photo.conds.file}{The name of a .csv file describing the photographic
  conditions under which each photograph was taken to produce the samples.
  Also, this argument can already be a \code{data.frame} with those photographic
  conditions.

  The following list is the set of columns expected in this data. It can
  contain more columns, but they will be ignored.

  \itemize{
    \item\code{crop.file.name}
    \item\code{lighting}
    \item\code{ISO}
    \item\code{shutter.speed}
    \item\code{aperture}
    \item\code{focal.length}
  }

  If your data contains these information but using another column names, you
  can map your column names to the expected ones using the option
  \code{\link{imgnoiser.option}('conds.col.map')}.

  For example, if you have a column named \code{ShutterSpeed} which is not
  the expected \code{shutter.speed} column name, the \code{conds.col.map}
  option must contain an entry with \code{ShutterSpeed} as key and
  \code{shutter.speed} as value, as in \code{list('ShutterSpeed' =
  'shutter.speed')}.

  The files named in the code{crop.file.name} column, regardless of their
  file name extension, must have the
  \href{http://en.wikipedia.org/wiki/FITS}{'FITS'} format.}

\item{crop.files.path}{The \code{photo.conds.file} argument refers to a file
  containing a column named \code{crop.file.name} with the name of each image
  sample file. Those files are expected to be in one same folder. This
  argument must contain the path to that folder, which should not be part of
  the \code{crop.file.name} values.}
}
\value{
The data resulting from this process can be seen using the functions
  \code{\link{hvdvm$get.var}}, \code{\link{hvdvm$get.cov}} and
  \code{\link{hvdvm$get.photo.conditions}}. Also, you can plot the result
  using the \code{\link{hvdvm$plot}} function.
}
\description{
Process the image samples with the \emph{hvdvm} technique
}
\details{
You can find more details about the input for this process in the
\emph{"Introduction to the *hvdvm* technique"} vignette.
}
\examples{
# In this example, the conditions file and the image sample files are in
# different folders, in this case specified with respect to the
# current working directory.
my.hvdvm$digest('./ISO100/White.csv', './ISO100/samples/')
}
\seealso{
See the section \emph{"Input data for the *hvdvm* Procedure"} in the
  \emph{"Introduction to the *hvdvm* technique"} vignette,
  \code{\link{hvdvm}}.
}

