% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$exists.model}
\alias{hvdvm$exists.model}
\alias{vvm$exists.model}
\title{Find if a model exists}
\arguments{
\item{model.name}{A character vector with the name of the models we want to
  know if already exists.}
}
\value{
A logical value. Returns TRUE only if the model given in
  \code{model.name} exists, otherwise returns false.
}
\description{
Find if a model with a given name has already been fitted.
}
\section{Usage}{

 \preformatted{
  hvdvm$exists.model(
     model.name = imgnoiser.option('fit.model.name')
     )

  vvm$exists.model(
     model.name = imgnoiser.option('fit.model.name')
     )
 }
}
\examples{
\dontrun{

# Check if there is a model named 'log-sqrt'
my.hvdvm$exists.model('log-sqrt')
#> [1] FALSE
# This function is "syntactic sugar" for the equivalent
('log-sqrt' \%in\% names(my.hvdvm$model.list))
}
}
\seealso{
\code{\link{hvdvm$fit.model}}, \code{\link{vvm$fit.model}}
}

