% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$fit.model}
\alias{hvdvm$fit.model}
\alias{vvm$fit.model}
\title{Fit a model to the noise data}
\arguments{
\item{model.name}{The name you assign to the model this function will
  create.

  You can build many models and use later this name to refer to any of them.
  You will use this name, for example, to get predicted values from the
  fitted model or to plot the data in the model or the predictions. If there
  is a model with this name, it will be over-written with the result of this
  new one.}

\item{model.family}{Name of one of the available models that will be fit to
  the data addressed by the \code{model.data} argument. If you use
  \code{NULL} for this argument, no model will be fitted over that data.

  The available model families are:

   \itemize{
     \item \code{lm} Linear models supported by the \code{stats::lm}
     function.

     \item \code{lmrob} MM-type estimators for linear models, supported by
     the \code{stats::lmrob} function.

     \item \code{smooth.spline} Cubic smoothing spline fitting, supported by
     the \code{stats::smooth.spline} function.
   }}

\item{degree}{The degree of a polynomial of \code{x} that will be fit over
  the data. This is assuming the model familly supports that kind of fitting,
  which is not the case for the \code{smooth.spline} model.}

\item{formula}{An optional formula to be used in the fitting proccess. If it
  is not provided, the formula will be constructed based on the \code{degree}
  argument. If a formula is provided the \code{degree} argument will be
  ignored.}

\item{model.data.name}{Identifies the specific data that will be part of the
  model.

  By default, there is only one model, named \emph{std.var} with variance data
  (from the hvdvm$var.df). This model has the \emph{mean} as predictor (x)
  and the \emph{var} as predicted (y) variable. Those variables are grouped
  by the corresponding \code{channel} variable (factor).

  You can add more models changing the option \code{get.model.src.data} using
  the \code{\link{imgnoiser.option}} function.}

\item{...}{Additional fitting model arguments that will be directly passed to
  function supporting the model referred by the \code{model.family}
  argument.

  By default, the functions supporting the model families are called with the
  model formula (or just \code{x} and \code{y} for the case of smooth.spline)
  as the only argument. All the additional arguments, supported by the
  function behind each model family, can be specified as part of this
  \code{...} argument, they are passed directly as received here to the
  corresponding model family function.}
}
\value{
The calling objects instance in invisible way (check the
  \code{\link{invisible}} function).

  This return value is irrelevant when the class is used in interactive
  way. It can be useful for developers extending the class.
}
\description{
Fits a model through specific variables in the data computed and collected by
the \code{digest} function. The arguments and semantics of this function are
the same in both, the \code{\link{hvdvm}} and the
\code{\link{vvm}} class.
}
\details{
The model concept bundles variables selected from those computed by the
\code{\link{hvdvm$digest}} or the \code{\link{vvm$digest}} function, to an
optional fitting model, that will be fitted over that selected data. Then
-for example- we can plot the data in the model with the predictions from the
fitted model.

All the parameters have sensible default values that you can customize as
options with \code{\link{imgnoiser.option}()} function.
}
\section{Usage}{

 \preformatted{
  hvdvm$fit.model(
     model.name = imgnoiser.option('fit.model.name'),
     model.family = imgnoiser.option('fit.model.family'),
     degree = 1L,
     formula = NULL,
     model.data.name = imgnoiser.option('fit.model.data'),
     ...
     )

  vvm$fit.model(
     model.name = imgnoiser.option('fit.model.name'),
     model.family = imgnoiser.option('fit.model.family'),
     degree = 2L,
     formula = NULL,
     model.data.name = imgnoiser.option('fit.model.data'),
     ...
     )
 }
}
\examples{
\dontrun{

# Fit a model using all default values.
my.hvdvm$fit.model()
#> The model was succesfully fitted using:
#> lm(formula = var ~ mean, data = model.src("std.var"))

# Fit a robust weighted model with the name 'weighted'
library(robustbase)
my.hvdvm$fit.model(model.name = 'weighted', model.family = 'lmrob', weights=1/mean^2)
#> The model was succesfully fitted using:
#> lmrob(formula = var ~ mean, data = model.src("std.var"), weights = 1/mean^2)
}
}
\seealso{
The \code{fit.model}, \code{get.model.predictions} and
  \code{get.model.src.data} options. The \code{hvdvm$digest()},
  \code{\link{vvm$digest}()} functions.
}

