% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/hvdvm_class.R
\name{hvdvm$get.cov}
\alias{hvdvm$get.cov}
\title{Get the resulting covariance data.}
\usage{
hvdvm$get.cov
}
\value{
A data.frame with the covariance data from the channels in each delta
  image computed by the \code{\link{digest}} function. This data frame
  contains the following columns:

\itemize{

 \item \code{cond}: A key for the photographic conditions common to the two
 processed samples (\code{pict1, pict2}). The detail of the photographic
 conditions is returned by the \code{\link{hvdvm$get.photo.conditions}}
 function, where the key is the value in this \code{cond} value.

 \item \code{pict1, pict2}: Name of the processed image samples files.

 \item \code{chan.a, chan.b}: The label of the channels whose covariance is shown
 in the column \code{cov}.

 \item \code{cov}: The half of the photosite values covariance in the
 \code{chan.a} and \code{chan.b} channels in the difference image, computed
 pixel by pixel, of the \code{pict1} image minus the \code{pict2} one.
}
}
\description{
Returns the covariance data computed by the \code{\link{digest}} function.
This values are half of the covariance in the delta image, computed for each
pair of channels combination.
}
\examples{
# print the head of the variance data frame
head(my.hvdvm$get.cov)

# Merge the variance and covariance data
merged <- merge(hvd$get.var, hvd$get.cov, by=c('pict1','pict2', 'cond'))
# Plot all the covariances (y axis) with respect to the means (x axis)
plot(merged$mean, merged$cov)
}
\seealso{
\code{\link{hvdvm$get.var}}, \code{\link{hvdvm$get.photo.conditions}}
}

