% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$get.model.predictions}
\alias{hvdvm$get.model.predictions}
\alias{vvm$get.model.predictions}
\title{Get the values predicted by a fitted model}
\arguments{
\item{model.name}{The name of the model whose predicitions are desired.}

\item{x}{The predictor values for which the predictions are desired. If this
  argument is omitted, the predictions for selected points in the range
  of the source data used to build the model are used.}

\item{select}{The channels for which we want to get the predictions.}

\item{conf.level}{The confidence level to get the predictions interval. It
  must be in the [0,1] range}
}
\value{
A data frame with the predictions from the model. In the following
  descriptions, we will call \code{pred.df} to this result.

   \itemize{
     \item \code{pred.df[,1]} The predictor variable in the data model. For
     the default model (\code{std.var} ), this corresponds to the
     \code{mean} column in the variance data.

     The range of values in this column spans the range of the original data
     used to fit the model. It contains 20 equally spaced values along the
     range of the source data, plus 10 values logarithmically equally spaced.
     If the range of the original data contains the zero or negative values,
     the predictor variable will contain 30 equally spaced values.

     \item \code{pred.df[,2]} The predicted variable according to the fitted
     model. For the \code{std.var} default model, this corresponds to the
     \code{var} column in the variance data.

     \item \code{pred.df[,3]} The variable used to group the prediitions. For
     the \code{std.var} default model, this corresponds to the
     \code{channel} column in the variance data.

     \item \code{fit.se} The prediction standard error.

     \item \code{lcl} The \emph{lower confidence limit}. It is computed as
     the predicted value minus the confidence factor multiplied by the
     fitting standard error.

     \item \code{ucl} The \emph{upper confidence limit}. It is computed as
     the predicted value plus the confidence factor multiplied by the fitting
     standard error.
   }
}
\description{
Get the values predicted from a given fitted model.
}
\section{Usage}{

 \preformatted{
  vvm$get.model.predictions(
     model.name = imgnoiser.option('fit.model.name'),
     x = NULL,
     select = NULL,
     conf.level,
     ...
     )

  hvdvm$get.model.predictions(
     model.name = imgnoiser.option('fit.model.name'),
     x = NULL,
     select = NULL,
     conf.level,
     ...
     )
 }
}
\examples{
\dontrun{

# Get the 'standard' model predictions
preds.df <- my.hvdvm$get.model.predictions()
# Set the lcl and ucl confidence limits using +/- 2.5 fit standard error
preds.df$lcl <- preds.df[,2L] - 2.5*preds.df$fit.se
preds.df$ucl <- preds.df[,2L] + 2.5*preds.df$fit.se
}
}
\seealso{
\code{\link{hvdvm$fit.model}}, \code{\link{vvm$fit.model}}
}

