% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/hvdvm_class.R
\name{hvdvm$get.photo.conditions}
\alias{hvdvm$get.photo.conditions}
\title{Get the distinct photographic conditions}
\usage{
hvdvm$get.photo.conditions
}
\value{
A data.frame containing the following columns:
 \enumerate{
    \item\code{cond}
     Is the key identifier of the conditions, which is referenced by the
     variance and the covariance data in the \code{\link{hvdvm$get.var}} and
     \code{\link{hvdvm$get.cov}} variables.

    \item\code{lighting}
    \item\code{ISO}
    \item\code{shutter.speed}
    \item\code{aperture}
    \item\code{focal.length}
    \item\code{count}
    The number of samples having the given conditions.
 }

This data prepared by the \code{\link{hvdvm$digest}} function selecting the
different combinations of values in the \code{.csv} input file. As the
\emph{hvdvm} technique requires pairs of images, only the conditions having
more than one image sample are included here.
}
\description{
Returns a code{data.frame} with information about the different photographic
conditions under which the photographs, used to get the samples, were taken.
It is computed by the \code{\link{hvdvm$digest}} function.
}
\examples{
# Find photographic conditions where more than 5 shots were taken
phcond <- my.hvdvm$get.photo.conditions()
subset(phcond, count > 5)
}
\seealso{
\code{\link{hvdvm$digest}}, \code{\link{hvdvm$get.cov}},
  \code{\link{hvdvm$get.var}}.

\code{\link{hvdvm$get.var}}, \code{\link{hvdvm$get.cov}}
}

