% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/hvdvm_class.R
\name{hvdvm$get.var}
\alias{hvdvm$get.var}
\title{Get the resulting variance data.}
\usage{
hvdvm$get.var
}
\value{
A data.frame with the variance data from the channels in the image
  samples processed by the \code{\link{digest}} function. It contains the
  following columns:

\itemize{

 \item \code{cond}: A key for the photographic conditions common to the two
 processed samples. The detail of the photographic conditions is returned
 by the \code{\link{hvdvm$get.photo.conditions}} function, where the key is
 the value in this \code{cond} value.

 \item \code{pict1, pict2}: Name of the processed image samples files.

 \item \code{channel}: The label of the channel whose processed mean
 and variance are shown in the following columns.

 \item \code{mean}: The mean of the photosites values of the average image,
 computed pixel by pixel, from the samples \code{pict1} and \code{pict2}.

 \item \code{var}: The half of the photosite values variance of the
 difference image, computed pixel by pixel, of the \code{pict1} image minus
 the \code{pict2} one.
}
}
\description{
Returns the variance data computed by the \code{\link{hvdvm$digest}}
function.
}
\examples{
\dontrun{
# print the head of the variance data frame
head(my.hvdvm$get.var)

# Pass the variance data to other variable
var <- my.hvdvm$get.var

# From here, it is just as like any data frame.
# You can, for example, get a subset with the red channel rows with mean above 2000
red.var <- subset(var, channel =='Red' & mean > 2000, var)
}
}
\seealso{
\code{\link{hvdvm$get.cov}}, \code{\link{hvdvm$get.photo.conditions}}
}

