% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$model.list}
\alias{hvdvm$model.list}
\alias{vvm$model.list}
\title{Get the list of models}
\value{
A list where the names are those given to the fitted models and the
  values are a character value with the call used to fit the model.
}
\description{
Get a list with the currently fitted models.
}
\section{Usage}{

 \preformatted{
  hvdvm$model.list

  vvm$model.list
 }
}
\examples{
\dontrun{

# Print the list of models
my.hvdvm$model.list
#> $standard
#> [1] "lm(formula = var ~ mean, data = model.src('std.var'))"
#>
#> $weighted
#> [1] "lmrob(formula = var ~ mean, data = model.src('std.var'), weights = 1/mean^2)"
}
}
\seealso{
\code{\link{hvdvm$fit.model}}, \code{\link{vvm$fit.model}}
}

