% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noiseVar_class.R
\name{hvdvm$new}
\alias{hvdvm$new}
\alias{vvm$new}
\title{Create a new class instance}
\usage{
hvdvm$new(
     channel.labels = imgnoiser.option('channel.labels'),
     green.channels = imgnoiser.option('green.channels'),
     has.RGGB.pattern = imgnoiser.option('has.RGGB.pattern'),
     avg.green.label = imgnoiser.option('avg.green.channel.label')
     )

  vvm$new(
     channel.labels = imgnoiser.option('channel.labels'),
     green.channels = imgnoiser.option('green.channels'),
     has.RGGB.pattern = imgnoiser.option('has.RGGB.pattern'),
     avg.green.label = imgnoiser.option('avg.green.channel.label')
     )
}
\arguments{
\item{channel.labels}{A vector of characters for the labeling of each one of
  the photosites colors in the (raw) image samples. Its default value is
  taken from \code{\link{imgnoiser.option}('channel.labels')}.}

\item{green.channels}{A vector with two integers, indices to the
  \code{channel.labels} vector, marking them as the green channels. By
  default is the value of \code{\link{imgnoiser.option}('green.channels')},
  which is \code{NA}}

\item{has.RGGB.pattern}{A logical value indicating if the image samples has
  the RGGB pattern in the photosites colors. Its default value is taken from
  \code{\link{imgnoiser.option}('has.RGGB.pattern')}, which by default is
  \code{FALSE}, meaning the values passed for the \code{channels.labels} and
  \code{green.channels} arguments are used as explained above.

  When it is \code{TRUE} the value for the \code{channels.labels} argument is
  taken from \code{\link{imgnoiser.option}('RGGB.channel.labels')} and the
  \code{green.channels} argument is set to \code{c(2L,3L)}. In this case the
  actual values passed for the \code{channels.labels} and
  \code{green.channels} arguments are silently ignored.}

\item{avg.green.label}{A single character value with the label for the
  averaged green channel. It's defaulr value is taken from
  \code{\link{imgnoiser.option}('avg.green.channel.label')}}
}
\value{
The new and initialized \code{hvdvm} class instance.
}
\description{
Creates and initializes a new hvdvm or vvm class instance. The arguments and
semantics of this function are the same in both classes. All arguments
have defaults based on option values, so you can set those option and call
this initialization functions without arguments.
}
\details{
With the \code{channel.labels} argument you can specify the labels for the
corresponding photistes color filters, starting at the top left corner, of
each of your samples. You will find an introduction to the topic, and a
procedure to identify the color of each channel, in the section \emph{"Sample
Borders and Channel Identification"} of the \emph{"Collecting Image Samples"}
vignette.

With the \code{green.channels} argument you can mark which of the labels in
the \code{channel.labels} vector correspond to the green channels. For
example, if you set the \code{channel.labels} as \code{c('Verde A', 'Azul',
'Rojo', 'Verde R')} (in spanish), you should mark the first and the last
channels ('verde' is the spanish word for green) as those corresponding to
the green channels. Knowing which are the green channels the class can
compute ditional statistics averaging both green channels.

If the photosites colors in your samples has the RGGB pattern, set to
\code{TRUE} the \code{has.RGGB.pattern} argument. In that case, the channel
labels are taken from the option \code{imgnoiser.RGGB.channel.labels}, which
by default contains the values \code{c('Red', 'Green R', 'Green B', 'Blue')}.
If you don't like those labels you can change the value of that option before
calling this function.
}
\examples{
# Create an instance accepting all the default argument values
my.hvdvm <- hvdvm$new()

# Create an instance for samples with the RGGB pattern
my.hvdvm <- hvdvm$new(is.RGGB = TRUE)

# Specify the channel labels in Italian and mark which of those
# are the green channels, both as options, and initialize the class
# without passing arguments.
imgnoiser.option('channel.labels', c('Verde', 'Blu', 'Rosso', 'Verde'))
imgnoiser.option('green.channels', c(1,4))
my.hvdvm <- hvdvm$new()
}
\seealso{
See the section \emph{"Sample Borders and Channel Identification"}
  in the \emph{"Collecting Image Samples"} vignette, \code{\link{hvdvm}},
  \code{\link{imgnoiser.option}}
}

