% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hvdvm-class.r
\name{hvdvm$photo.conditions.df}
\alias{hvdvm$photo.conditions.df}
\title{Get the distinct photographic conditions}
\value{
A data.frame containing the following columns:
 \enumerate{
    \item\code{cond}
     Is the key identifier of the conditions, which is referenced by the
     variance and the covariance data in the \code{\link{hvdvm$var.df}} and
     \code{\link{hvdvm$cov.df}} variables.

    \item\code{lighting}
    \item\code{ISO}
    \item\code{shutter.speed}
    \item\code{aperture}
    \item\code{focal.length}
    \item\code{count}
    The number of samples having the given conditions.
 }

This data prepared by the \code{\link{hvdvm$digest}} function selecting the
different combinations of values in the \code{.csv} input file. As the
\emph{hvdvm} technique requires pairs of images, only the conditions having
more than one image sample are included here.
}
\description{
Returns a code{data.frame} with information about the different photographic
conditions under which the photographs, used to get the samples, were taken.
It is computed by the \code{\link{hvdvm$digest}} function.
}
\section{Usage}{

 \preformatted{
  hvdvm$photo.conditions.df
 }
}
\examples{
\dontrun{

# Find photographic conditions where more than 5 shots were taken
phcond <- my.hvdvm$photo.conditions.df
subset(phcond, count > 5)
}
}
\seealso{
\code{\link{hvdvm$digest}}, \code{\link{hvdvm$cov.df}},
  \code{\link{hvdvm$var.df}}.

\code{\link{hvdvm$var.df}}, \code{\link{hvdvm$cov.df}}
}

