% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noiseVar_class.R
\name{hvdvm$remove.model}
\alias{hvdvm$remove.model}
\alias{vvm$remove.model}
\title{Remove a fitted model}
\usage{
hvdvm$remove.model(
     model.name = stop('A model name argument is required.')
     )

  vvm$remove.model(
     model.name = stop('A model name argument is required.')
     )
}
\arguments{
\item{model.name}{The name of the model that will be removed (deleted).}
}
\description{
Remove a model fitted with a given name.
}
\examples{
\dontrun{

my.hvdvm$remove.model('weighted')
#> The model with name "standard" has been removed.
}
}

