\name{hvdvm-class}
\docType{class}
\alias{hvdvm}
\title{Analysis of raw image noise using the 'hvdvm' technique.}
\description{
The \code{hvdvm} class contains tools for the measurement, analysis and study
and of the half variance and covariance of delta images for the study of
noise in raw image files.

You can read the \emph{"\code{hvdvm} Class Introduction"} vignette for an
introduction to the subject and the \emph{"\code{hvdvm} and \code{vvm} vvm
Class Usage"} vignette for additional details about the general
workflow using this class.
}
