% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/imgnoiser.R
\docType{package}
\name{imgnoiser-package}
\alias{imgnoiser-package}
\title{The imgnoiser package}
\description{
Tools to measure, analyze and profile raw image noise.
}
\details{
The \pkg{imgnoiser} package contains tools aimed to automatize
the measurement, analysis and profiling of image noise.
}

