% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fit-model.r
\name{imgnoiser.get.model.source.data}
\alias{imgnoiser.get.model.source.data}
\title{Returns a model data}
\usage{
imgnoiser.get.model.source.data(
     model.name,
     noise.obj
     )
}
\arguments{
\item{model.name}{The name of the required data. Actually, the package only
  supports the data model named 'standard' which is the default value for
  this argument, see details bout this in the description of the return
  value.}

\item{noise.obj}{An instance, either of the class \code{\link{hvdvm}}
  or \code{\link{vvm}}. This object provide the data from which the
  data model will be constructed.}
}
\value{
A list with the following elements:
 \itemize{

   \item \code{data} A data frame with the following variables:

     \itemize{
         \item \code{x} Numeric values. They will be plotted along the x axis
         and will be used as predictor variable for model fitting.

         \item \code{y} Numeric values. They will be plotted along the y axis
         and will be used as the predicted variable for model fitting.

         \item \code{split.by} A factor. This variable is optional. If present
         it will be used to split the (x, y) values in as many groups as
         different values are in this variable.
     }

   \item \code{confid.factor} A numeric value: the confidence factor that
   will be used for computing the confidence limits around the fitted values.
   For example 1.96 corresponding to a 97.5\% confidence level for fitting
   errors with a normal distribution. In this case the limits are the fitted
   values +/- 1.96 times the standard error.

   \item \code{label} A list with the labels for the elements in the data
   model and for the its representation in a plot.

     \itemize{
         \item \code{term} A named vector with the names corresponding to the
         variables in the data frame.

         The names of the elements in the vector must be
         \code{c('x','y','split.by')}, and as values the corresponding labels
         for those variables, for example \code{c('Mean','Var','Channel')},
         so the value of \code{term['x']} is \code{'Mean'}, the label for the
         'x' variable.

         \item \code{xlab} The title for the x axis in the plot.
         \item \code{ylab} The title for the y axis in the plot.
         \item \code{main} The main plot title.
     }
 }

 For the \emph{standard} model, which actually is the only one supported, the
 data comes from the variance information in the instance object given in the
 \code{noise.obj} argument, get through the use of the function
 \code{noise.obj$var.df}.

 The \code{x} and \code{y} values in the function return value are the
 \code{mean} and the \code{var} variables in the variance data, and as
 \code{split.by} is used the \code{channel} variable.

 The \emph{confidence factor} (\code{confid.factor}) value is taken from the
 \code{confid.factor} option in \code{\link{imgnoiser.option}}.

 The labels (\code{xlab}, \code{ylab}) and title (\code{main}) are customized
 according the class of the \code{noise.obj} argument.
}
\description{
Returns the data corresponding to a given model name. It is used internally
and is exposed only to provide to the user a easy way to extend it adding
new data models.
}
\details{
This function is internally called by the \code{fit.model} function in the
classes \code{hvdvm} and \code{vvm}. It is not inteded to be called directly.
It is made public to allow the user extend it providing other data models for
the noise analysis.

This function is the default value of the \code{get.model.src.data} option.
The value of that option can be changed to refer to another function, with
the same signature as this, providing additional data models. This other
function might call this one for the data models already implemented here.
}
\seealso{
\code{\link{hvdvm$fit.model}}, \code{\link{vvm$fit.model}}, the
  option \code{get.model.src.data} of \code{\link{imgnoiser.option}},
  \code{\link{imgnoiser.fit.model}},
  \code{\link{imgnoiser.model.predictions}}
}

