% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/persistence.r
\name{imgnoiser.load}
\alias{imgnoiser.load}
\title{Load imoiser objects}
\usage{
imgnoiser.load(
     file.name    = stop("A 'file.name' argument is required."),
     load.options = imgnoiser.option('obj.load.options')
  )
}
\arguments{
\item{file.name}{Name of the file containing the object that will be loaded.
  This is file previously saved with the \code{\link{imgnoiser.save}}
  function.}

\item{load.options}{If TRUE and if the the \code{imgnoiser} options were
  saved in the file, those options will be retrieved from the file and set as
  the current ones.}
}
\value{
The object read from the \code{file.name} file.
}
\description{
Load an instance of a imgnoiser object saved in a specified file. An instance
of the \code{imgnoiser} classes \code{\link{hvdvm}},
\code{\link{vvm}} is loaded from the a given file.
}
\examples{
\dontrun{

# Retrieve the my.hvdvm object previously saved with imgnoiser.save
my.hvdvm <- imgnoiser.load('my.hvdvm')
#> An instance of the "hvdvm" was succesfully read from the "my.hvdvm.imn" file.
}
}
\seealso{
\code{\link{imgnoiser.save}}
}

