% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.r
\name{imgnoiser.options.force.reset}
\alias{imgnoiser.options.force.reset}
\title{Force the reset of options}
\usage{
imgnoiser.options.force.reset()
}
\description{
Reset all options to their default values.
}
\details{
By default, the package installation only set the package options missing in
the current set of R options. However sometimes is required to reset all
options.
}

