% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.r
\name{imgnoiser.options.load}
\alias{imgnoiser.options.load}
\title{Load the package options}
\usage{
imgnoiser.options.load(
     file.name    = stop("A 'file.name' argument is required.")
     )
}
\arguments{
\item{file.name}{Name of the file containing the options that will be loaded.
  This is file previously saved with the \code{\link{imgnoiser.save}}
  function.}
}
\description{
Load the package options from a file.
}
\examples{
\dontrun{

# Load 'my.options'
imgnoiser.options.load('my.options')
}
}
\seealso{
\code{\link{imgnoiser.options.save}}, \code{\link{imgnoiser.save}},
  \code{\link{imgnoiser.load}}
}

