% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/options.r
\name{imgnoiser.options.save}
\alias{imgnoiser.options.save}
\title{Save the package options}
\usage{
imgnoiser.options.save(
     file.name = stop("A 'file.name' argument is required."),
     stop.overwrite = imgnoiser.option('stop.save.overwrite')
     )
}
\arguments{
\item{file.name}{The name of the file where the options will be saved. If
  this file name does not already contains '.imn' as file name extension, it
  will be appended to it.}

\item{stop.overwrite}{If TRUE will only allow to save the options in a new
  file. The function will not overwrite a file an already existing
  \code{file.name} file. If FALSE, will silently overwrite a file with the
  same name if it already exists.}
}
\description{
Save all the package options in a file.
}
\examples{
# Save into 'my.options' file.
imgnoiser.options.save('my.options')
}
\seealso{
\code{\link{imgnoiser.options.load}}, \code{\link{imgnoiser.save}},
  \code{\link{imgnoiser.load}}
}

