% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/persistence.r
\name{imgnoiser.save}
\alias{imgnoiser.save}
\title{Save imoiser objects}
\usage{
imgnoiser.save(
     obj            = stop("An 'obj' argument is required."),
     file.name      = deparse(substitute(obj)),
     save.options   = imgnoiser.option('obj.save.options'),
     stop.overwrite = imgnoiser.option('stop.save.overwrite')
   )
}
\arguments{
\item{obj}{The \code{imgnoiser} object that will be saved.}

\item{file.name}{Name of the file where the \code{obj} will be saved. If this
  argument does not contains '.imn' as file name extension, it will be
  appended to it.}

\item{save.options}{If TRUE all the \code{imgnoiser} options and the \code{obj}
  object will be saved together in the \code{file.name} file.}

\item{stop.overwrite}{If TRUE will only allow to save the \code{obj} in a new
  file. The function will not overwrite a file an already existing
  \code{file.name} file. If FALSE, will silently overwrite a file with the
  same name if it already exists.}
}
\description{
Save an instance of a \code{imgnoiser} object to the specified file. An
instance of the classes \code{\link{hvdvm}}, \code{\link{vvm}}
with all its internal data and models is saved to a file with a given name.
}
\examples{
\dontrun{

# Save the my.hvdvm object, with his own name "my.hvdvm"
imgnoiser.save(my.hvdvm)
#> The "my.hvdvm" object was succesully saved in the "my.hvdvm.imn" file.

# Try to save it again, with imgnoiser.option('stop.save.overwrite') set to FALSE
imgnoiser.save(my.hvdvm)
#> Error: The file "my.hvdvm.imn" already exists.

# Force the overwriting
imgnoiser.save(my.hvdvm, stop.overwrite = FALSE)
#> The "my.hvdvm" object was succesully saved in the "my.hvdvm.imn" file.

# Save once again, this time with a given file name
imgnoiser.save(my.hvdvm, "my.hdvm.backup")
#> The "my.hvdvm" object was succesully saved in the "my.hdvm.backup.imn" file.
}
}
\seealso{
\code{\link{imgnoiser.load}}
}

