% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{L.Lab.from.adobeRGB}
\alias{L.Lab.from.adobeRGB}
\title{Get L.Lab tonal values from Adobe RGB ones}
\usage{
L.Lab.from.adobeRGB(adobeRGB)
}
\arguments{
\item{Adobe}{RGB tonal values in \code{[0, 1]}}
}
\value{
L.Lab tonal values in \code{[0, 100]}
}
\description{
Convert Adobe RGB tonal values to Lightness (L from Lab) ones.
}

