% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{L.Lab.from.linear}
\alias{L.Lab.from.linear}
\title{Get L.Lab tonal values from linear ones}
\usage{
L.Lab.from.linear(linear)
}
\arguments{
\item{linear}{tonal values in \code{[0, 1]}}
}
\value{
L.Lab tonal values in \code{[0, 100]}
}
\description{
Convert linear tonal values to Lighteness (L in Lab).
}

