% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{L.Lab.from.sRGB}
\alias{L.Lab.from.sRGB}
\title{Get L.Lab tonal values from sRGB ones}
\usage{
L.Lab.from.sRGB(sRGB)
}
\arguments{
\item{sRGB}{tonal values in \code{[0, 1]}}
}
\value{
L.Lab tonal values in \code{[0, 100]}
}
\description{
Convert sRGB tonal values to Lightness (L from Lab) ones.
}

