% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{linear.from.adobeRGB}
\alias{linear.from.adobeRGB}
\title{Get linear tonal values from Adobe RGB ones}
\usage{
linear.from.adobeRGB(adobeRGB)
}
\arguments{
\item{Adobe}{RGB tonal values in \code{[0, 1]}}
}
\value{
linear tonal values in \code{[0, 1]}
}
\description{
Convert Adobe RGB tonal values to linear ones.
}

