% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{linear.from.L.Lab}
\alias{linear.from.L.Lab}
\title{Get linear tonal values from L.Lab ones}
\usage{
linear.from.L.Lab(L.Lab)
}
\arguments{
\item{L.Lab}{tonal values in \code{[0, 100]}}
}
\value{
linear tonal values in \code{[0, 1]}
}
\description{
Convert Lighteness (L in Lab) tonal values to linear ones.
}

