% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{linear.from.proPhoto}
\alias{linear.from.proPhoto}
\title{Get linear tonal values to proPhoto ones}
\usage{
linear.from.proPhoto(proPhoto)
}
\arguments{
\item{proPhoto}{tonal values in \code{[0, 1]}}
}
\value{
linear tonal values in \code{[0, 1]}
}
\description{
Convert linear tonal values to proPhoto ones.
}

