% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{linear.from.sRGB}
\alias{linear.from.sRGB}
\title{Get linear tonal values from sRGB ones}
\usage{
linear.from.sRGB(sRGB)
}
\arguments{
\item{sRGB}{tonal values in \code{[0, 1]}}
}
\value{
linear tonal values in \code{[0, 1]}
}
\description{
Convert sRGB tonal values to linear ones.
}

