% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{proPhoto.from.linear}
\alias{proPhoto.from.linear}
\title{Get proPhoto tonal values from linear ones}
\usage{
proPhoto.from.linear(linear)
}
\arguments{
\item{linear}{linear tonal values in \code{[0, 1]}}
}
\value{
proPhoto tonal values in \code{[0, 1]}
}
\description{
Convert linear tonal values to proPhoto ones.
}

