% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imnr-utils.r
\name{select.files}
\alias{select.files}
\title{Query the files in a given folder}
\usage{
select.files(from, to, extension, path)
}
\arguments{
\item{from}{}

\item{to}{The \code{from} and \code{to} are the alphabetical range of the
  desired files in folder given in \code{path}. These names should not
  include the name extension which is specified with \code{extension}.}

\item{extension}{The file name extension common to the desired files.}
}
\value{
A character vector with the file names satisfying the selection.
}
\description{
Return a vector with the names of the files in the folder \code{path}, whose
names are alphabetically between \code{from} and \code{to} and have the name
extension given in \code{extension}.
}
\examples{
\dontrun{

# Select the file in the folder './samples' whose name is between
# _DSC5695.NEF and _DSC5716.NEF
files <- select.files('_DSC5695', '_DSC5716', '.NEF', './samples')
}
}

