% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/split-cfa.r
\name{split_channels}
\alias{split_channels}
\title{Split a raw image to get the four channels}
\usage{
split_channels(file.name, file.path)
}
\arguments{
\item{file.name}{The file name of the raw image. It must have the extension
'.fit' or '.pg'.}

\item{file.path}{The path to the file referred in \code{file.name}}
}
\value{
An list of four matrices with the image from each channel.
}
\description{
Split a raw image to get the four channels
}

