% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{sRGB.from.adobeRGB}
\alias{sRGB.from.adobeRGB}
\title{Get sRGB tonal values from Adobe RGB ones}
\usage{
sRGB.from.adobeRGB(adobeRGB)
}
\arguments{
\item{sRGB}{tonal values in \code{[0, 1]}}
}
\value{
Adobe RGB tonal values in \code{[0, 1]}
}
\description{
Convert Adobe RGB tonal values to sRGB ones.
}

