% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{sRGB.from.L.Lab}
\alias{sRGB.from.L.Lab}
\title{Get sRGB tonal values from L.Lab ones}
\usage{
sRGB.from.L.Lab(L)
}
\arguments{
\item{L.Lab}{tonal values in \code{[0, 100]}}
}
\value{
sRGB tonal values in \code{[0, 1]}
}
\description{
Convert Lightness (L from Lab) tonal values to sRGB ones.
}

