% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color-utils.r
\name{sRGB.from.linear}
\alias{sRGB.from.linear}
\title{Get sRGB from linear tonal values}
\usage{
sRGB.from.linear(linear)
}
\arguments{
\item{linear}{tonal values in \code{[0, 1]}}
}
\value{
sRGB tonal values in \code{[0, 1]}
}
\description{
Convert linear tonal values to sRGB.
}

