% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vvm-class.r
\name{vvm$cov.df}
\alias{vvm$cov.df}
\title{Get the resulting noise covariance data}
\value{
A data.frame with the covariance data from the channels in the image
  samples computed by the \code{\link{vvm$digest}} function. It contains the
  following columns:

\itemize{

 \item \code{pict1}: Name of the processed image sample file.

 \item \code{chan.a, chan.b}: The label of the channels whose covariance is shown
 in the \code{cov} column.

 \item \code{cov}: The covariance between the channels \code{chan.a, chan.b} in the
 \code{pict1} image sample.
}
}
\description{
Returns the covariance data computed by the \code{\link{vvm$digest}} function
from each images sample.
}
\section{Usage}{

 \preformatted{
  vvm$cov.df
 }
}
\examples{
\dontrun{

# print the head of the variance data frame
head(my.vvm$cov.df)

# Merge the variance and covariance data
merged <- merge(hvd$var.df, hvd$cov.df, by=c('pict1','pict2', 'cond'))
# Plot all the covariances (y axis) with respect to the means (x axis)
plot(merged$mean, merged$cov)
}
}

