% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/vvm_class.R
\name{vvm$get.cov}
\alias{vvm$get.cov}
\title{Get the resulting covariance data.}
\usage{
vvm$get.cov()
}
\value{
A data.frame with the covariance data from the channels in the image
  samples computed by the \code{\link{digest}} function. It contains the
  following columns:

\itemize{

 \item \code{cond}: A key for the photographic conditions common to the two
 processed samples. The detail of the photographic conditions is returned
 by the \code{\link{vvm$get.photo.conditions}} function, where the key is
 the value in this \code{cond} value.

 \item \code{pict1, pict2}: Name of the processed image samples files.

 \item \code{chan.a, chan.b}: The label of the channels whose covariance is shown
 in the following column \code{cov}.

 \item \code{cov}: The covariance between the channels \code{pict1, pict2} of the
 difference image, computed pixel by pixel, of the \code{pict1} image minus
 the \code{pict2} one.
}
}
\description{
Returns the covariance data computed by the \code{\link{digest}} function
from each pair of images samples taken under identical photographic
conditions.
}
\examples{
# print the head of the variance data frame
head(my.vvm$get.cov())

# Merge the variance and covariance data
merged <- merge(hvd$get.var(), hvd$get.cov(), by=c('pict1','pict2', 'cond'))
# Plot all the covariances (y axis) with respect to the means (x axis)
plot(merged$mean, merged$cov)
}

