% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vvm-class.r
\name{vvm$var.df}
\alias{vvm$var.df}
\title{Get the resulting variance data.}
\value{
A data.frame with the variance data from the channels in the image
  samples processed by the \code{\link{vvm$digest}} function. It contains the
  following columns:

\itemize{

 \item \code{pict}: Name of the processed image sample file.

 \item \code{channel}: The label of the channel in the sample file that was
 processed.

 \item \code{mean}: The mean of the photosites values in the channel
 \code{channel} in the image sample \code{pict}.

 \item \code{var}: The variance of the photosites values in the channel
 \code{channel} in the image sample \code{pict}.
}

As a convenience, when possible, there is a channel labeled \emph{"Green
Avg"} with information summarizing both green channels in the sample. However
this label can be customized in the instance object creation with the
\code{\link{vvm$new}} function.

When this data is included the \code{mean} column contains the mean
photosites values considering both green channels as a single whole, the same
applies for the corresponding \code{var} column.

This data is computed only if the green channels have been specified in the
instance creation using the \code{green.channels} argument of the
\code{\link{vvm$new}} function.
}
\description{
Returns the variance data computed by the \code{\link{vvm$digest}} function
from each pair of images samples taken under identical photographic
conditions.
}
\section{Usage}{

 \preformatted{
  vvm$var.df
 }
}
\examples{
\dontrun{

# print the head of the variance data frame
head(my.vvm$var.df)

# Pass the variance data to other variable
var <- my.vvm$var.df

# From here, it is just as like any data frame.
# You can, for example, get a subset with the red channel rows with mean above 2000
red.var <- subset(var, channel =='Red' & mean > 2000, var)
 }
}

