\name{vvm-class}
\docType{class}
\alias{vvm}
\title{Measurement and study of raw image noise.}
\description{
The \code{vvm} class contains tools for the measurement, analysis and study of
the mean, variance and covariance of pixel values in raw image files.

You can read the \emph{"\code{vvm} Class Introduction"} vignette for an
introduction to the subject and the \emph{"\code{hvdvm} and \code{vvm} vvm
Class Usage"} vignette for additional details about the general
workflow using this class.
}
