% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$channel.labels}
\alias{hvdvm$channel.labels}
\alias{vvm$channel.labels}
\title{Get the channel labels}
\value{
A character vector with five values. The first two are for the first
  row in the 2x2 Bayer pattern, the second pair are for the second row, and
  the fifth to label computations merging both green channels.
}
\description{
A vector with the channel labels. This labels are set when an instance of the
class is constructed using the \code{new} function.
}
\section{Usage}{

 \preformatted{
  hvdvm$channel.labels

  vvm$channel.labels
 }
}
\examples{
\dontrun{

# Print the channel labels
my.hvdvm$channel.labels
#> [1] "Blue"    "Green B" "Green R" "Red"
}
}
\seealso{
\code{\link{hvdvm$new}}, \code{\link{vvm$new}}
}

