% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$get.model}
\alias{hvdvm$get.model}
\alias{vvm$get.model}
\title{Get a model objects}
\value{
A list whose names are the channel labels and the values are the
  objects fitted for each channel.
}
\description{
Returns a list with the objects fitted in a model with a
given name.
}
\details{
The R functions that actually fit the models, like lm() or lmrob(), return an
object which can be used in other R functions supporting them. As one model
is fitted to the data from each channel there are four channel models for
each single model fitted in call to the \code{fit.model} function. This
function returns a list of those channel model objects.
}
\section{Usage}{

 \preformatted{
  hvdvm$get.model(
     model.name = imgnoiser.option('fit.model.name')
     )

  hvdvm$get.model(
     model.name = imgnoiser.option('fit.model.name')
     )
 }
}
\examples{
\dontrun{

# Get a list with the 'weighted' model fitting objects
ch.models <- my.hvdvm$get.model('weighted')
#> class(ch.models)
#> [1] "list"
#> names(ch.models)
#> [1] "Blue"    "Green B" "Green R" "Red"
}
}

