% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noise-class.r
\name{hvdvm$remove.model}
\alias{hvdvm$remove.model}
\alias{vvm$remove.model}
\title{Remove a fitted model}
\arguments{
\item{model.name}{The name of the model that will be removed (deleted).}
}
\description{
Remove a model fitted with a given name.
}
\section{Usage}{

 \preformatted{
  hvdvm$remove.model(
     model.name = stop('A model name argument is required.')
     )

  vvm$remove.model(
     model.name = stop('A model name argument is required.')
     )
 }
}
\examples{
\dontrun{

my.hvdvm$remove.model('weighted')
#> The model with name "standard" has been removed.
}
}

