% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vvm-class.r
\name{vvm$digest}
\alias{vvm$digest}
\title{Process the raw image samples.}
\arguments{
\item{file.name.from}{}

\item{file.name.to}{The \code{file.name.from} and \code{file.name.to} are the
  alphabetical range of the desired files from the folder given in
  \code{path.to.files}. These names should not include the file name
  extension, which is specified with \code{file.name.ext}.}

\item{path.to.files}{The file path where the raw image file samples are located.}

\item{file.name.ext}{The file name extension of the raw image file samples.
  It can be 'fit', 'fits' or 'pgm'. This file name extension may start or not
  with a period as in '.fit' or fit'.

  The supported values for the \code{img.file.name.ext} are '.fit' or '.fits'
  for the \href{http://fits.gsfc.nasa.gov/}{'FITS'} format or '.pgm' for the
  \href{http://netpbm.sourceforge.net/doc/pgm.html}{'PGM'} format.}
}
\value{
A \code{invisible} instance of the calling object.
}
\description{
Read raw image samples and compute variance and covariance statistics for
each channnel.
}
\details{
The calling arguments describe the list of raw image files that will be
processed. The image samples file names are expected to have the following
pattern:

\emph{\code{path.to.files}/\code{<img.file.base>}.\code{<file.name.ext>}}

Where \code{<img.file.base>} is alfabetically between the
\code{file.name.from} and \code{file.name.to} arguments.

In other words, the folder given in \code{path.to.files} is scanned looking
for files with base name between (and including) \code{file.name.from} and
\code{file.name.to}, having all of them the extension \code{file.name.ext}.

If your sample file names do not follow this pattern, you can specify them as
a character vector as argument for the \code{<file.name.from>} parameter.

The \code{\link{select.file.range}} may help you to get a baseline selection with
the files you want to get processed by this function.
}
\section{Usage}{

 \preformatted{
  vvm$digest(
     file.name.from = stop("The 'file.name.from' argument is missing."),
     file.name.to   = stop("The 'file.name.to' argument is missing."),
     path.to.files  = './',
     file.name.ext  = '.pgm'
     )
 }
}
\examples{
\dontrun{

# Create a new vvm instance
my.vvm <- vvm$new(has.RGGB.pattern=TRUE)

img.path <- 'D:/Noise-Study/Nikon-D7000/ISO-100'
my.vvm$digest('_ODL5695', '_ODL5767', '.pgm', img.path)
}
}
\seealso{
See the section \emph{"Input data for the *vvm* Procedure"} in the
  \emph{"Introduction to the *vvm* technique"} vignette,
  \code{\link{vvm}}.
}

