% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/vvm_class.R
\name{vvm$digest}
\alias{vvm$digest}
\title{Process the image samples.}
\usage{
vvm.doc$digest(
     img.file.name = stop("The 'img.file.name' argument is required"),
     img.file.count = NULL,
     file.path = './',
     img.file.name.ext = '.fit'
)
}
\arguments{
\item{img.file.name}{A single character value, with the prefix of the raw
  image samples file names. It can also be a character vector with the whole
  file name sequence of the image samples, but without the file path, which
  should be given in the \code{file.path} argument.}

\item{img.file.count}{The number of image sample files. If the
  \code{<img.file.name>} argument contains the whole file name sequence this
  argument is irrelevant and consequently ignored.}

\item{file.path}{The file path where the raw image file samples are located.}

\item{img.file.name.ext}{The file name extension of the raw image file
  samples. It can be 'fit' or 'fits'. This file name extension may start with
  a period as in '.fit' or not ('fit'). When required, the function will add
  the period '.' to build the sequence of file names.}
}
\value{
A \code{invisible} instance of the calling object.
}
\description{
Read image samples and compute variance and covariance statistics for
each channnel.
}
\details{
The calling arguments describe the sequence of raw image files that will be
processed. The image samples file names are expected to have the following
pattern:

\emph{\code{<img.file.name>}<index>\code{<img.file.name.ext>}}

Where \code{<img.file.name>} and \code{<img.file.name.ext>} are arguments of
this function. An example of file names with this pattern is ('crop_1.fit',
'crop_2.fit', ...), where 'crop_' is given in \code{<img.file.name>} and
'fit' is given in \code{<img.file.name.ext>}.

The file path to the files is given in the \code{file.path} argument. i.e. it
is must not be part of the \code{<img.file.name>} argument.

The numbering iof the file names must start with 1 and end with
\code{img.file.count} value without any gap in in the sequence.

If your sample file names do not follow this pattern, you can specify them as
a character vector used as argument for the \code{<img.file.name>} parameter.
}
\examples{
\dontrun{

# Create a new vvm instance
my.vvm <- vvm$new(has.RGGB.pattern=TRUE)

# Process 40 files names 'crop_1.fit', 'crop_2.fit'...
# in the ./data-raw/ folder
my.vvm$digest("crop_", 40, './data-raw/')
}
}
\seealso{
See the section \emph{"Input data for the *vvm* Procedure"} in the
  \emph{"Introduction to the *vvm* technique"} vignette,
  \code{\link{vvm}}.
}

